some refactoring of client and db functions
