+++ univ-boilerplate/db/index.js
@@ -1,42 +1,7 @@
-// Connect to DB
-const { Client } = require('pg');
-const DB_NAME = 'localhost:5432/grace'
-const DB_URL = process.env.DATABASE_URL || `postgres://${ DB_NAME }`;
-const client = new Client(DB_URL);
-// const {getAllProducts, getProductById, updateProductById, deleteProductById} = require('./products');
 
-// database methods
-async function createUser({username, password, email, address}) {
-  try{
-      const {rows: [user]} = await client.query(`
-          INSERT INTO users(username, password, email, address)
-          VALUES ($1, $2, $3, $4)
-          ON CONFLICT (email,username) DO NOTHING
-          RETURNING *;
-      `, [username, password, email, address]);
+const client = require('./client')
 
-      return user
-
-  } catch(err){
-      throw err;
-  }
-}
-
-async function createProduct({name, price, description, image}){
-  try{
-    const {rows: [product]} = await client.query(`
-      INSERT INTO products(name, price, description, image)
-      VALUES ($1, $2, $3, $4)
-      RETURNING *;
-    `, [name, price, description, image])
-
-    return product
-
-  }catch(err){
-    throw err
-  }
-}
-
+//does not have its own js file yet
 async function createOrder({userId}){
   try{
     const {rows: [order]} = await client.query(`
@@ -52,6 +17,7 @@ async function createOrder({userId}){
   }
 }
 
+//does not have its own js file yet
 async function createReview({productId, userId, rating, content}) {
   try{
     const {rows: [review]} = await client.query(`
@@ -67,6 +33,7 @@ async function createReview({productId, userId, rating, content}) {
   }
 }
 
+//does not have its own js file yet
 async function createOrderProducts({orderId, productId}) {
   try{
     const {rows: [orderProducts]} = await client.query(`
@@ -82,29 +49,14 @@ async function createOrderProducts({orderId, productId}) {
   }
 }
 
-async function createUsersProducts({userId, productId, quantity}) {
-  try{
-    const {rows: [usersProducts]} = await client.query(`
-      INSERT INTO users_products("userId", "productId", quantity)
-      VALUES ($1, $2, $3)
-      RETURNING *;
-    `, [userId, productId, quantity])
-
-    return usersProducts
-
-  } catch(err){
-    throw err
-  }
-}
-
 // export
 module.exports = {
   client,
-  createUser,
-  createProduct,
   createOrder,
   createReview,
   createOrderProducts,
-  createUsersProducts,
+  ...require('./users'),
+  ...require('./products'),
+  ...require('./users_products'),
   // db methods
 }
