+++ univ-boilerplate/db/users.js
@@ -1,4 +1,4 @@
-const {client} = require("./index");
+const client = require("./client");
 const bcrypt = require('bcrypt');
 
 async function createUser({username, password, email, address}) {
