+++ univ-boilerplate/db/users_products.js
@@ -1,4 +1,4 @@
-const { client } = require('./index')
+const client = require('./client')
 
 async function createUsersProducts({userId, productId, quantity}) {
     try{
